refactor: rename useSlotContextMenu to slotMenuService
Not a composable — returns no reactive state and has no lifecycle
integration.

Addresses review feedback:
https://github.com/Comfy-Org/ComfyUI_frontend/pull/9045#discussion_r2925519398
