+++ src/renderer/extensions/vueNodes/composables/slotMenuService.test.ts
@@ -26,7 +26,7 @@ import {
   hasAnySlotAction,
   removeSlot,
   renameSlot
-} from './useSlotContextMenu'
+} from './slotMenuService'
 
 describe(canRenameSlot, () => {
   beforeEach(() => {
